Set default action for SIGPIPE in child
If xterm is started from xwin-xdg-menu, minimized test case:

$ (/bin/echo A; /bin/echo B) | head -1
/bin/echo: write error: Broken pipe

However this does not happen if xterm is started from mintty window by
'xterm -display :0'.

It seems that SIGPIPE is handled differently. I guess SIGPIPE is handled
by SIG_IGN if the terminal started from xwin-xdg-menu, otherwize, it is
handled by SIG_DFL.

See:

https://cygwin.com/pipermail/cygwin/2019-August/242060.html
https://cygwin.com/pipermail/cygwin-developers/2021-September/012389.html

As discussed, it's probably ok to just do this for SIGPIPE, as that's
the only one gtk sets to ignored, but let's do it for all signals, just
in case we ever set another one similarly.